Initial commit
Adding source files and images to work with. BMP image read and write is
implemented using QDBMP librtary. The top file contains preliminar
implementation of the "blocker" (separates the image in a block vector
to operate with) "imager" (truns back block vectors into images after
operation) a configurable mask creator and rotation and iversion
routines for the blocks. The code works only in 8x8 blocks and 512x512
images as of now.